Move UserVariable dialogs in separate file and to src subdir. On non batch start initialise the user var GUI, else leave the default
